+++ Assets/ActionThrowStone.cs
@@ -3,7 +3,9 @@ using System.Collections.Generic;
 using UnityEngine;
 using UnityEngine.Networking;
 
-public class ActionThrowStone : NetworkBehaviour, IAction {
+public class ActionThrowStone : Action, IAction {
+  public GameObject stonePrefab;
+
   public int TpCost(){
     return(30);
   }
@@ -25,13 +27,18 @@ public class ActionThrowStone : NetworkBehaviour, IAction {
   }
 
   public void BeginAction(GameObject targetObject){
-    //if(cursor.standingUnit){
-      //cursor.standingUnit.ReceiveDamage(5);
-    //}
+    Cursor cursor = targetObject.GetComponent<Cursor>();
+    GameObject stoneObject = Instantiate(stonePrefab, Unit().transform.position, Quaternion.identity);
+    Stone stone = stoneObject.transform.Find("Stone").GetComponent<Stone>();
+    stone.action = this;
+    stone.cursor = cursor;
   }
 
   public void DoAction(Cursor cursor){
-
+    if(cursor.standingUnit){
+      cursor.standingUnit.ReceiveDamage(5);
+    }
+    Unit().FinishAction();
   }
 
   public bool NeedsLineOfSight(){